fix(bar,histogram): complete v3 validation and selection fixes
Preserve source-index selection and exact histogram bins, unify range and layout calculations, and render validation errors without throwing.

Correct grouped style defaults and numeric precision, add focused regression coverage, and update the migration guide. Keep the animation test removed.

Verify production and test-source compilation, the full lint gate, and lightweight JVM helper and validation tests. Leave UI and platform test execution to CI.

+++ sample/app/src/commonMain/kotlin/dev/hdcode/charts/app/demo/bar/BarDemo.kt
@@ -117,8 +117,8 @@ fun BarChartDemo(viewModel: BarChartViewModel = koinViewModel()) {
                                         .single()
                                         .values.size,
                                 useBarColors = true,
-                                minValue = controlsState.minValue.toFloat(),
-                                maxValue = controlsState.maxValue.toFloat(),
+                                minValue = controlsState.minValue.toDouble(),
+                                maxValue = controlsState.maxValue.toDouble(),
                             ),
                     )
                 }
